pyproject.toml
[tool.poetry]
name = "projeto backand capyba"
version = "0.1.0"
description = ""
authors = ["Nycolas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
gunicorn = "^23.0.0"
uvicorn = "^0.31.1"
waitress = "^3.0.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.28.0"
pytest-django = "^4.9.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
